+++ hs.js
@@ -2,6 +2,8 @@ var core = require('./lib/core')
 var base = require('./lib/base')
 var list = require('./lib/list')
 
+list.n = require('./lib/list.n')
+
 exports.core = core
 exports.base = base
 exports.list = list